Implement a preliminary partitioning scheme.
In this scheme, a box has 6 surfaces: front, back, top, bottom,
left and right. A box is specified by two diagonal points at the
left front bottom and right back top corners. There are 12 edges
which are the intersection of two adjacent surfaces such as front
and top. A box is authoritative of the halo regions inside it. A
point can be in at most 3 halo surfaces, and at most 3 halo edges.
It could also not be in any of the halo regions.